move none tests class to seperated examples package
