+++ services/src/test/java/io/scalecube/services/examples/GreetingRequest.java
@@ -1,4 +1,4 @@
-package io.scalecube.services;
+package io.scalecube.services.examples;
 
 import java.time.Duration;
