+++ services/src/test/java/io/scalecube/services/examples/GreetingService.java
@@ -1,4 +1,4 @@
-package io.scalecube.services;
+package io.scalecube.services.examples;
 
 import io.scalecube.services.annotations.Service;
 import io.scalecube.services.annotations.ServiceMethod;
