+++ services/src/test/java/io/scalecube/services/routings/examples/CanaryService.java
@@ -1,9 +1,9 @@
 package io.scalecube.services.routings.examples;
 
-import io.scalecube.services.GreetingRequest;
-import io.scalecube.services.GreetingResponse;
 import io.scalecube.services.annotations.Service;
 import io.scalecube.services.annotations.ServiceMethod;
+import io.scalecube.services.examples.GreetingRequest;
+import io.scalecube.services.examples.GreetingResponse;
 
 import reactor.core.publisher.Mono;
