+++ services/src/test/java/io/scalecube/services/routings/examples/DummyRouter.java
@@ -1,7 +1,8 @@
-package io.scalecube.services;
+package io.scalecube.services.routings.examples;
 
 import static com.google.common.base.Preconditions.checkArgument;
 
+import io.scalecube.services.ServiceReference;
 import io.scalecube.services.api.ServiceMessage;
 import io.scalecube.services.registry.api.ServiceRegistry;
 import io.scalecube.services.routing.Router;
